Optimize map reversal with extension method
Refactored the duplicated map reversal logic into a reusable extension
method 'swapKV', streamlining the process of generating spline series in
the chart widget. This change not only reduces redundancy but also
enhances readability and maintainability of the codebase. Additionally,
cast operation is added to ensure type safety when accessing map values
as numerics.